Added server connection between front and backend to actually request API updates within the server.
Added a background image because recently learned
how some consider it better practice to create
the design the content is supposed to fall in
before actually putting the content in. So
background image plus styles.css to touch that up.
And then aside from the weather script and the css
I also added the weather component in the HTML and
some filler text for the news just to see how it
would look. And then there's the weather script
to actually display the server code on the
frontend.